Tested with local hotplate_consumer project
find_package(hotplate) worked if I configured with CMAKE_PREFIX_PATH set to the directory containing hotplateConfig.cmake.

I did need to make the change here for this to work. I commented the set_and_check lines, because I don't know if the variables are necessary now that we have IMPORTED targets.